Broke GenericAction out into separate components. Put all monitoring actions in a separate folder
