+++ app/views/patients/show.html.erb
@@ -13,7 +13,7 @@
 <% if current_user.can_modify_subject_status? %>
 <div class="card mx-2 mt-4 mb-4 card-square">
   <h5 class="card-header">Monitoring Actions</h5>
-  <%= react_component('subject/MonitoringActions', {
+  <%= react_component('subject/monitoring_actions/MonitoringActions', {
                         current_user: current_user,
                         user_can_transfer: current_user.can_transfer_patients?,
                         authenticity_token: form_authenticity_token,
